Added `projected` and `infimum` ...
... for partition power states to the main `Network.Legion` export list. These
operate on the partition power state, and there purpose is to allow
implementers of the `Persistence` layer some insight into what their
persistence layers are actually storing. This enables the user to, for
instance, maintain some kind of index on the persistence layer.

The thing is, it is the user who controls the type of the partition state,
which is wrapped by Legion in a partition power state. However the partition power
state is opaque to the user, so they have no insight into what is being stored
even though they control the type of what is being stored. Exporting these two
functions gives the user some limited insight into what is being stored,
without exposing the internals of the partition power state.