Optimize the migration code. ...
The way we have to do a full listing of the persistence layer is bad
enough, and we definitely have plans in the future to fix that.

In the meantime though, we at least don't want to do that full listing
even if we know there is no migration to run.

This change simply avoids doing the listing if we know there are no
keys that need migrating.