Fixed runtime peer awareness bug. ...
There was a problem where the connection manager part of the
runtime system was not picking up peers that had joined the cluster.
Specifically, if knowledge about a peer was obtained via gossip, instead
of that peer directly joining us, then the connection manager wouldn't
pick it up. This is because the state machine failed inform the runtime
whenever it received a cluster state update from another peer.

We probably need to change the way the runtime system detects changes
in the state; probably by inspecting the state directly instead of have
the state machine communicate directly with the runtime.