Tweaked the node state debug JSON ...
... to show structured information about the propStates.

This change has a side effect that prevents us from showing JSON-structured
'ClusterState' data like we had been before, because that would force us to
expose a `ToJSON s` constraint to the user. Making the user implement 'ToJSON'
on their state type is just cruel and sadistic, and we aren't going to do it.