Use `-XConstraintKinds` for `LegionConstraints` ...
... instead of created a whole new typeclass. This basically allows us
to create type aliases for a set of constraints, which is what we were
trying to do in the first place with `LegionConstraints`.

The advantage here is that error message related to these constraints
are much more specific. So a user knows that they are missing, say, the
`Binary` instance for their data type, instead of just `LegionConstraints`
as a whole.