refactor - Make first class `Runtime` data type. ...
Instead of having `forkLegionary` return an unnamed function that allows
access to the runtime system, we instead wrap that access function in
its own data type. This is mainly to make it easier later on when we
extend the number of ways the user can access the runtime system (for
instance with index lookups).