Don't allow state inspection. ...
Specifically, don't export the `projected` and `infimum` functions
which allow the user-provided persistence layer to inspect the state
of the partition. I am uneasy about allowing this type of capability
because of its potential for abuse. The only reason I had exported
them in the first place was to allow for the user-implemented indexing
experiment implemented in legion-cache, that was a precursor to true
Legion indexing.

Now that Legion indexing exists, there is no compelling reason to
provide this kind of state inspection.