Removed `PartitionKey` from the signature of `handleRequest`. ...
Having the partition key available as an input to the request handler has
turned out to be completely useless in every case so far.  The request
handler is responsible for implementing the application business logic,
whereas the partition key is strictly a Legion concern. Therefore, in
addition to being useless, it is probably actually a very bad thing to
allow the business logic to have access to this information.  Therefore,
I am removing it from the interface.