Unified request handler. ...
The request handling logic, which used to be split between `handleRequest`
and the `ApplyDelta` typeclass has now been unified into the `ApplyDelta`
typeclass. This is better for a lot of reasons including reduced
interface complexity. But it is particularly useful because it is the
first step in implementing the option to provide fully consistent (but
not partition-tolerant) request responses.