Fixed functional dependencies on ApplyDelta. ...
ApplyDelta is supposed to represent a way of encoding enumerated function
types that map `s -> (o, s)`. Even though everything compiled fine the
way we had it, making sure that both `s` and `o` are fully specified
by the `i` type more closely embodies the spirit that `i` represents a
monomorphic function.

it is more in the spirit of having `i`
represent some set of functions to make sure th