Formalize the DifferentOrigins exception. ...
Instead of just using a descriptive String error for merge failures,
formalize the error as a first-class Exception. At the moment, there
is only one type of error condition that can result from a merge, and
that is the case where the two power states do not share the same
origin.