Add infimum outputs to the PowerState module. ...
An "infimum output" is the output of an event that has reached the
infimum. Normally, we achieve eventual state consistency by possibly
replaying events gathered from different peers in the correct order,
but by that time the event output has already been sent to the
user. Therefore, while the partition state has an eventual consistency
guarantee, the output for each individual event does not.

Once an event has reached the infimum, however, its order relative to
other events is fixed for all time. Therefore its output at that time
is strongly consistent. This change provides a mechanism for accessing
the strongly consistent outputs for events that have become fixed behind
the infimum.