Fixed some space leaks. ...
Two of the "leaks" were not, strictly speaking, leaks. They were the
result of thunk buildups that would eventually release the memory whenever
certain capabilities of the framework were exercised. For these we want
to enforce strict evaluation because there is no guarantee that the user
would choose to exercise those capabilities in a timely fashion, if ever.

The third one was a legitimate leak caused by the use of
`Control.Monad.forever`, where "legitimate" means that the space was
not recoverable under any conditions. Still not entirely sure what was
going on there. See the comments in the code.